py/misc.h: Fix missing cast in mp_clz_mpi.
gcc version of __builtin_clzl and __builtin_clzll return int,
So they should be casted to uint32_t.

+++ py/misc.h
@@ -356,7 +356,7 @@ static inline uint32_t mp_clzll(unsigned long long x) {
 // Microsoft don't ship _BitScanReverse64 on Win32, so emulate it
 static inline uint32_t mp_clzll(unsigned long long x) {
     unsigned long h = x >> 32;
-    return h ? mp_clzl(h) : (mp_clzl(x) + 32);
+    return h ? mp_clzl(h) : (mp_clzl((unsigned long)x) + 32);
 }
 #endif
 
@@ -384,9 +384,9 @@ static inline uint32_t mp_clz_mpi(mp_int_t x) {
 
     // ugly, but should compile to single intrinsic unless O0 is set
     if (mp_check(sizeof(mp_int_t) == sizeof(long))) {
-        return mp_clzl((unsigned long)x);
+        return (uint32_t)mp_clzl((unsigned long)x);
     } else {
-        return mp_clzll((unsigned long long)x);
+        return (uint32_t)mp_clzll((unsigned long long)x);
     }
 }
